+++ src/Command/Right.java
@@ -24,7 +24,7 @@ public class Right implements Command {
 
 
     @Override
-    public int returnValue() {
-        return (int)degree;
+    public double returnValue() {
+        return (double)degree;
     }
 }
